+++ pkg/ui/react-app/src/thanos/Navbar.tsx
@@ -28,6 +28,7 @@ interface NavDropDown {
 const navConfig: { [component: string]: (NavConfig | NavDropDown)[] } = {
   query: [
     { name: 'Graph', uri: '/new/graph' },
+    { name: 'Stores', uri: '/new/stores' },
     { name: 'Status', children: [{ name: 'Command-Line Flags', uri: '/new/flags' }] },
   ],
 };