+++ pkg/ui/ui.go
@@ -34,6 +34,7 @@ var (
 		"/targets",
 		"/tsdb-status",
 		"/version",
+		"/stores",
 	}
 )